Renaming pyfun to chebpy
In the initial stage of development, the working name for the project
was "pyfun". Now the name has been changed to "chebpy". The idea here
is that, since Matlab Chebfun is extension of Matlab (to continuous
functions and operators as per [Battles and Trefethen (2004)]), chebpy
can be seen as a corresponding extension of numpy.