Adding tests for Chebtech.__div__
Was necessary to be a little careful here not to test the __div__
operation for demoninators with roots, since this would result in
singlular behaviour and non-convergence of the result. As such an
additional boolean flag has been added to the tests/utilities.py
testfunctions variable indicating whether the function has roots
in the unit interval. This flag is used to determine whether or not
to add the __div__ test when cycling over the different pairwise
combinations of the elements of testfunctions.

Since testfunctions is utilised elsewhere in the test suite, it was
necessary to add dummy _ variables wherever these were unpacked.