Modifiying testfunctions collection
Such that the third argument (True/False) corresponds to whether or not
there are roots present anywhere on the real line. This change was made
so that this collection of test functions can be used (filtered by the
Boolean flag) to test the functionality of higher-level classes such as
Bndfun and Chebfun.